refactor: split API types by features
